keyboard + new SW
keys 4, 5, 1, 2 on the numpad do the buttons.
after noticing that the rolling release version of the sw appease to
download all files every time, I've opted for a static almost offline
only version, relying on a list of all needed files in a versioned
cache.
When I update the game, I chenge the name of the cache.